pyproject.toml
[project]
name = "llm-as-judge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.4",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "pandas>=2.3.3",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
]
